Moved files around and split them up so common stuff is in Arr
