+++ arr/__init__.py
@@ -1,8 +1,8 @@
 """Asynchronous Python client for Sonarr."""
-from arr.exceptions import (  # noqa
+from .exceptions import (  # noqa
     ArrAccessRestricted,
     ArrConnectionError,
     ArrError,
     ArrResourceNotFound,
 )
-from .sonarr import Sonarr  # noqa
+from .client import Client
