+++ arr/__version__.py
@@ -1,0 +1,3 @@
+"""Asynchronous Python client for Arr."""
+
+__version__ = "0.4.0"
