+++ arr/const.py
@@ -1,0 +1,1 @@
+"""Constants for Arr."""
