--- sonarr/__version__.py
@@ -1,3 +1,0 @@
-"""Asynchronous Python client for Sonarr."""
-
-__version__ = "0.3.0"
